# pkg/openapi3/URL_types.gen.go
// Package openapi3 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.1 DO NOT EDIT.
package openapi3

// URL defines model for URL.
type URL struct {
	HTMLVersion            *string `json:"HTMLVersion,omitempty"`
	HaveLoginForm          *bool   `json:"HaveLoginForm,omitempty"`
	HeadingsCount          *string `json:"headingsCount,omitempty"`
	Id                     *string `json:"id,omitempty"`
	InaccessibleLinksCount *int32  `json:"inaccessibleLinksCount,omitempty"`
	LinksCount             *int32  `json:"linksCount,omitempty"`
	PageTitle              *string `json:"pageTitle,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error *string `json:"error,omitempty"`
}

// ReadURLsResponse defines model for ReadURLsResponse.
type ReadURLsResponse struct {
	URL *URL `json:"URL,omitempty"`
}

// SearchURLsResponse defines model for SearchURLsResponse.
type SearchURLsResponse struct {
	URL *URL `json:"URL,omitempty"`
}

// SearchURLsRequest defines model for SearchURLsRequest.
type SearchURLsRequest struct {
	URL *string `json:"URL,omitempty"`
}

// CreateURLJSONRequestBody defines body for CreateURL for application/json ContentType.
type CreateURLJSONRequestBody SearchURLsRequest